// src/features/projects/types/project.ts
export interface ProjectPhase {
    id: string;
    name: string;
    completed: boolean;
}

export interface Project {
    id: string; // Database PK
    folio?: string; // PROY-2026-... (generated by backend)
    title: string;
    estado: string;
    municipio: string;
    categoria: string;
    tipoEscala: string;
    capas: string[];
    fases: ProjectPhase[];
    avance: number;
    status: 'Completado' | 'En proceso' | 'Pendiente';
}